# api/openapi.yaml
openapi: 3.1.0
info:
  title: ISC Core API
  version: 0.1.0
  description: |
    ISC（接入编排器）内核的本地管理接口。

    ## 设计约束

    - **仅本机可达**：内核只监听命名管道 / Unix socket / 回环，绝不对外暴露。
      所有请求（含回环）都必须携带 `Authorization: Bearer <token>`，
      token 从数据目录下的 `runtime.json` 读取。见 docs/DECISIONS.md D09。
    - **快同步 + 慢异步**：列表与详情查询同步返回；所有可能耗时的操作
      （证书签发、凭据校验、防火墙应用、外部可达性验证）返回 `202` 与
      `job_id`，进度通过 `GET /v1/jobs/{id}` 或 WebSocket 事件流获取。
      见 docs/DECISIONS.md D08。
    - **契约优先**：本文件是接口的唯一真理。Go 服务端接口由 oapi-codegen 生成，
      前端类型由 openapi-typescript 生成。修改接口必须先改本文件。

    ## M0 阶段范围

    当前仅有元信息、健康检查、事件流与任务引擎。业务端点随里程碑加入。
servers:
  - url: http://localhost
    description: 实际地址由运行时的 runtime.json 决定（命名管道 / Unix socket / 回环）
...
security:
  - bearerAuth: []
paths:
...
  /v1/meta:
    get:
      tags: [Meta]
      operationId: getMeta
      summary: 元信息与平台能力
      description: |
        返回版本、构建信息、运行平台，以及各平台后端（防火墙 / 服务管理 /
        网卡监控 / 密钥库 / 传输 / 低端口绑定）的实现状态。

        下游 GUI 应当依据 `capabilities` 决定是否置灰相关功能，
        而不是自己去猜 `runtime.GOOS`。
      responses:
        "200":
          description: 元信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: 任务列表
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          required: false
          description: 按状态过滤，可重复传入。
          schema:
            type: array
            items:
              $ref: "#/components/schemas/JobStatus"
          style: form
          explode: true
        - name: kind
          in: query
          required: false
          description: 按任务类型过滤。
          schema:
            type: string
      responses:
        "200":
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/jobs/{id}/cancel:
    post:
      tags: [Jobs]
      operationId: cancelJob
      summary: 取消任务
      description: |
        请求取消一个尚未结束的任务。取消是协作式的：任务实现需自行响应
        `context.Context` 的取消信号。对已结束的任务调用返回 `409`。
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: 已受理取消请求（任务可能仍处于 canceling 状态）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/debug/noop:
    post:
      tags: [Debug]
      operationId: runNoopJob
      summary: 提交一个空转任务（仅用于管线自检）
      description: |
        **非产品接口。** 提交一个按步进上报进度的空转任务，用于验证
        任务引擎与事件流是否正常。计划在 M1 之后移除。

        `steps` 控制步数，`step_ms` 控制每步耗时。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoopRequest"
      responses:
        "202":
          description: 任务已提交
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        default:
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------------------
  # DNS 服务商元信息
  # ---------------------------------------------------------------------------
...
  /v1/settings:
    get:
      tags: [Settings]
      operationId: getSettings
      summary: 读取运行时设置
      responses:
        "200":
          description: 设置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        default:
          $ref: "#/components/responses/Problem"

    patch:
      tags: [Settings]
      operationId: updateSettings
      summary: 修改运行时设置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsPatch"
      responses:
        "200":
          description: 修改后的完整设置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------------------
  # 审计
  # ---------------------------------------------------------------------------
  /v1/audit:
    get:
      tags: [Audit]
      operationId: listAudit
      summary: 查询审计日志
      description: |
        记录所有**写操作**（谁、什么时候、对什么、做了什么、结果如何）。

        存在的意义：内核以系统服务身份运行、能改防火墙、能重写整个 DNS 区域。
        出问题时"最后一次改动是什么、谁改的"是排查的第一入口。
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: action
          in: query
          required: false
          description: 按动作过滤。
          schema:
            type: string
        - name: result
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AuditResult"
      responses:
        "200":
          description: 审计记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
        default:
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------------------
  # 配置导入导出
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        token 位于数据目录的 `runtime.json` 中（字段 `token`），
        由内核首次启动时生成的 32 字节随机值经 base64url 编码得到。
  parameters:
    JobId:
      name: id
      in: path
      required: true
      description: 任务 ID
      schema:
        type: string
        minLength: 1
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 分页游标，取自上一次响应的 next_cursor。
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 单页条数。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: 错误响应（RFC 9457 problem+json）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 定义的错误对象。
      required: [type, title, status]
      properties:
        type:
          type: string
          description: 标识错误类型的 URI。
          examples: ["https://isc.local/problems/not-found"]
        title:
          type: string
          description: 简短的人类可读摘要（已本地化）。
        status:
          type: integer
          description: HTTP 状态码。
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: 本次错误的具体说明（已本地化）。
        instance:
          type: string
          description: 出错的具体请求路径。
        code:
          type: string
          description: |
            稳定的机器可读错误码，供客户端分支判断。
            与 title/detail 不同，该字段不随语言变化。
          examples: ["job_not_found", "job_not_cancelable"]
        request_id:
          type: string
          description: 请求 ID，便于在内核日志中定位。

    # ------------------------------------------------------------------
    # 元信息
    # ------------------------------------------------------------------
...
    Meta:
      type: object
      required: [version, api_version, os, arch, started_at, capabilities]
      properties:
        version:
          type: string
          description: 内核版本，开发构建为 dev。
        api_version:
          type: string
          description: 本接口的版本，当前为 v1。
        commit:
          type: string
          description: 构建时的 git commit，未知时为空串。
        build_time:
          type: string
          description: 构建时间（RFC 3339），未知时为空串。
        os:
          type: string
          description: 运行平台（GOOS）。
          examples: [windows, linux, darwin]
        arch:
          type: string
          description: 运行架构（GOARCH）。
          examples: [amd64, arm64]
        started_at:
          type: string
          format: date-time
        capabilities:
          $ref: "#/components/schemas/Capabilities"
    Capabilities:
      type: object
      description: 各平台后端的实现状态。不可用的后端意味着对应功能降级为引导模式。
      required: [firewall, service_manager, ip_monitor, secret_store, transport, low_port_binder]
      properties:
        firewall:
          $ref: "#/components/schemas/ImplState"
        service_manager:
          $ref: "#/components/schemas/ImplState"
        ip_monitor:
          $ref: "#/components/schemas/ImplState"
        secret_store:
          $ref: "#/components/schemas/ImplState"
        transport:
          $ref: "#/components/schemas/ImplState"
        low_port_binder:
          $ref: "#/components/schemas/ImplState"
    ImplState:
      type: object
      required: [available, backend]
      properties:
        available:
          type: boolean
          description: 为 false 时该功能降级为引导模式（只提示用户该做什么）。
        backend:
          type: string
          description: 后端名称。
          examples: ["nftables", "netsh", "unsupported:firewall"]
        note:
          type: string
          description: 给用户看的补充说明（已本地化）。

    # ------------------------------------------------------------------
    # 任务
    # ------------------------------------------------------------------
    JobStatus:
      type: string
      enum:
        - pending
        - running
        - succeeded
        - failed
        - canceled
      description: |
        pending  已提交，尚未开始
        running  执行中
        succeeded 成功结束
        failed   失败结束（见 error 字段）
        canceled 被取消
    Job:
      type: object
      required: [id, kind, status, progress, created_at]
      properties:
        id:
          type: string
        kind:
          type: string
          description: 任务类型标识，供客户端分类展示。
          examples: ["debug.noop", "credentials.verify", "certs.issue"]
        status:
          $ref: "#/components/schemas/JobStatus"
        progress:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: 进度，0 到 1。未知进度时为 0。
        message:
          type: string
          description: 当前步骤的人类可读描述（已本地化）。
        result:
          description: 任务成功时的结果，结构由任务类型决定。
        error:
          $ref: "#/components/schemas/Problem"
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    JobAccepted:
      type: object
      required: [job_id]
      properties:
        job_id:
          type: string
    JobList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        next_cursor:
          type: string
          description: 为空表示没有更多数据。
    NoopRequest:
      type: object
      properties:
        steps:
          type: integer
          minimum: 1
          maximum: 100
          default: 5
        step_ms:
          type: integer
          minimum: 0
          maximum: 10000
          default: 200
        fail_at_step:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
          description: 大于 0 时在第 N 步失败，用于验证失败路径。

    # ------------------------------------------------------------------
    # DNS 服务商
    # ------------------------------------------------------------------
...
    Settings:
      type: object
      required: [lang, log_level]
      properties:
        lang:
          type: string
          enum: [zh-CN, en]
        log_level:
          type: string
          enum: [debug, info, warn, error]
        event_buffer_size:
          type: integer
          minimum: 100
          maximum: 100000
          description: 事件环形缓冲容量。调大可延长断线补发的覆盖窗口。
        notify_on_ip_change:
          type: boolean
          description: 地址或 IPv6 前缀变化时是否发送通知（M4 起生效）。
    SettingsPatch:
      type: object
      description: 只提交需要修改的字段。
      properties:
        lang:
          type: string
          enum: [zh-CN, en]
        log_level:
          type: string
          enum: [debug, info, warn, error]
        event_buffer_size:
          type: integer
          minimum: 100
          maximum: 100000
        notify_on_ip_change:
          type: boolean

    # ------------------------------------------------------------------
    # 审计
    # ------------------------------------------------------------------
    AuditResult:
      type: string
      enum: [success, failure, denied]
    AuditEntry:
      type: object
      required: [id, ts, action, result]
      properties:
        id:
          type: integer
          format: int64
        ts:
          type: string
          format: date-time
        action:
          type: string
          description: 稳定的机器可读动作名。
          examples: ["credential.create", "credential.update", "config.import"]
        target:
          type: string
          description: 操作对象的标识（例如凭据 ID 或名称）。
        result:
          $ref: "#/components/schemas/AuditResult"
        detail:
          type: string
          description: 补充说明（已本地化，**不含敏感值**）。
        request_id:
          type: string
        remote:
          type: string
          description: 来源描述。本机管理通道下通常是传输类型而非 IP。
    AuditList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditEntry"
        next_cursor:
          type: string

    # ------------------------------------------------------------------
    # 配置导入导出
    # ------------------------------------------------------------------
...